[Fix] Responsibility of classes related to AttackController
