+++ Assets/Scripts/Attack/AttackController.cs
@@ -4,33 +4,250 @@ using UnityEngine;
 using UnityEngine.UI;
 using System.Linq;
 
-/// <summary>
-/// 制約：このクラスは、
-/// BaseAttackController
-/// SingleAttackController
-/// RangeAttackController
-/// の3つのクラスがアタッチされているGameObjectにアタッチすること。
-/// </summary>
-public class AttackController : MonoBehaviour
+namespace AC
 {
-	[SerializeField]
-	private Map _map;
-	[SerializeField]
-	private Units _units;
-
-	private BaseAttackController _bac;
-	private SingleAttackController _sac;
-	private RangeAttackController _rac;
-
-	private void Start()
+	public class BaseAttackController
 	{
-		_bac = gameObject.GetComponent<BaseAttackController>();
-		_sac = gameObject.GetComponent<SingleAttackController>();
-		_rac = gameObject.GetComponent<RangeAttackController>();
+		// ==========変数==========
+
+		
+		private DamageCalculator _dc;
+
+
+		// ==========関数==========
+
+
+		public BaseAttackController(DamageCalculator dc)
+		{
+			_dc = dc;
+		}
+
+		/// <summary>
+		/// 特定のマスの敵を攻撃する
+		/// </summary>
+		public void AttackToUnit(Unit attacker, Unit defender, Attack attack)
+		{
+			// BattleSceneに移動してバトルをする (取り敢えず要らない)
+			// Battle_SceneController.attacker = attacker;
+			// Battle_SceneController.defender = defender;
+			// BattleSceneに移動.
+			// SceneManager.LoadScene("Battle", LoadSceneMode.Additive);
+
+			// ダメージ計算を行う
+			int damage = _dc.CalcurateDamage(attacker, attack, defender, defender.Floor);
+
+			// ダメージを適用する
+			defender.Damage(damage);
+		}
+	}
+
+
+	public class SingleAttackController
+	{
+		// ==========変数==========
+		
+
+		private Map _map;
+		private Units _units;
+		private BaseAttackController _bac;
+
+
+		// ==========関数==========
+
+
+		public SingleAttackController(Map map,Units units,BaseAttackController bac)
+		{
+			_map = map;
+			_units = units;
+			_bac = bac;
+		}
+
+		/// <summary>
+		/// 攻撃可能なマスをハイライトする。
+		/// </summary>
+		/// <returns>攻撃対象が居るか否か(コマンド選択可否に使うのかな？)</returns>
+		public bool SetAttackableHighlight(Unit attacker, SingleAttack attack)
+		{
+			var hasTarget = false;
+			Floor startFloor = attacker.Floor;
+			foreach (var floor in _map.GetFloorsByDistance(startFloor, attack.RangeMin, attack.RangeMax))
+			{
+				if (floor.Unit == null || floor.Unit.Belonging == attacker.Belonging) continue;
+
+				// 取り出したマスにユニットが存在し, そのユニットが敵軍である場合
+				hasTarget = true;
+				floor.SetAttackableHighlight();
+			}
+			return hasTarget;
+		}
+
+		/// <summary>
+		/// 対象ユニットに攻撃
+		/// </summary>
+		/// <param name="target">攻撃先(マス座標)</param>
+		/// <returns>ユニットがあるかどうか</returns>
+		public bool Attack(Unit attacker, Vector2Int target, Attack attack)
+		{
+			var defender = _units.GetUnit(target.x, target.y);
+
+			if (defender == null) return false;
+
+			_bac.AttackToUnit(attacker, defender, attack);
+
+			_map.ClearHighlight();
+
+			_units.FocusingUnit.IsMoved = true;
+
+			return true;
+		}
+	}
+
+
+	public class RangeAttackController
+	{
+		// ==========変数==========
+		
+
+		private Map _map;
+		private Units _units;
+		private BaseAttackController _bac;
+
+
+		// ==========関数==========
+
+
+		public RangeAttackController(Map map,Units units, BaseAttackController bac)
+		{
+			_map = map;
+			_units = units;
+			_bac = bac;
+		}
+
+		/// <summary>
+		/// 攻撃可能範囲を検索する
+		/// </summary>
+		private List<Vector2Int> GetAttackable(Unit unit, Attack attack, int attackDir)
+		{
+			// sinRot = sin(attackDir * PI/2)  (cosRotも同様) 
+			int sinRot = (attackDir % 2 == 0) ? 0 : (2 - attackDir);
+			int cosRot = (attackDir % 2 == 1) ? 0 : (1 - attackDir);
+
+			// attacker's place
+			int cx = unit.X;
+			int cy = unit.Y;
+
+			//wikipedia,回転行列を参照
+			var attackables = attack.Range.Select(p => new Vector2Int(
+				p.x * cosRot - p.y * sinRot + cx,
+				p.x * sinRot + p.y * cosRot + cy
+				)).ToList();
+			return attackables;
+		}
+
+		/// <summary>
+		/// 特定の位置にあるマスに攻撃可能ハイライトを点ける
+		/// </summary>
+		private void SetAttackableHighlight(List<Vector2Int> attackables)
+		{
+			// この関数を呼び出すとき、"必ず"ハイライトを1度全て解除するはず。
+			_map.ClearHighlight();
+
+			foreach (var attackable in attackables)
+			{
+				var floor = _map.GetFloor(attackable.x, attackable.y);
+				if (floor != null) floor.SetAttackableHighlight();
+			}
+		}
+
+		/// <summary>
+		/// 攻撃する方角を変更します（可能なら）
+		/// </summary>
+		/// <param name="befDir">先程まで向いていた方角</param>
+		/// <param name="isClockwise">押されたボタンが時計回りか否か</param>
+		/// <returns>今から見る方角</returns>
+		public int UpdateAttackableHighlight(Unit attacker, RangeAttack attack, int befDir, bool isClockwise)
+		{
+			// 回転できない場合は、その場で終了
+			if (!attack.IsRotatable) return befDir;
+
+			// 回転させる
+			int nowDir = (befDir + (isClockwise ? 3 : 1)) % 4;
+
+			// 攻撃範囲を計算する
+			var attackables = GetAttackable(attacker, attack, nowDir);
+
+			SetAttackableHighlight(attackables);
+
+			return nowDir;
+		}
+
+		/// <summary>
+		/// 攻撃可能ハイライトを初期設定する
+		/// </summary>
+		public int InitializeAttackableHighlight(Unit attacker, RangeAttack attack)
+		{
+			// 初期方角 : 陣営によって初期方角を変えるならここを変える
+			int startAttackDir = 0;
+
+			var attackables = GetAttackable(attacker, attack, startAttackDir);
+
+			SetAttackableHighlight(attackables);
+
+			return startAttackDir;
+		}
+
+		/// <summary>
+		/// 範囲内に居るユニットに攻撃
+		/// (範囲攻撃の赤マス選択時に呼び出される)
+		/// </summary>
+		/// <returns>範囲内に、敵が1体でも居たかどうか</returns>
+		public bool Attack(Unit attacker, Attack attack)
+		{
+			bool unitExist = false;
+			var attackRanges = _map.GetAttackableFloors();
+
+			// 攻撃した範囲全てに対して、
+			foreach (var attackRange in attackRanges)
+			{
+				// 敵Unitの存在判定を行い、
+				var defender = _units.GetUnit(attackRange.X, attackRange.Y);
+				if (defender == null) continue;
+				unitExist = true;
+				_bac.AttackToUnit(attacker, defender, attack);
+			}
+
+			_map.ClearHighlight();
+			_units.FocusingUnit.IsMoved = true;
+			return unitExist;
+		}
+	}
+
+}
+
+
+public class AttackController
+{ 
+	private Map _map;
+	private Units _units;
+	private DamageCalculator _dc;
+
+	private AC.BaseAttackController _bac;
+	private AC.SingleAttackController _sac;
+	private AC.RangeAttackController _rac;
+
+	public AttackController(Map map,Units units,DamageCalculator dc)
+	{
+		_map = map;
+		_units = units;
+		_dc = dc;
+
+		_bac = new AC.BaseAttackController(_dc);
+		_sac = new AC.SingleAttackController(_map, _units, _bac);
+		_rac = new AC.RangeAttackController(_map, _units, _bac);
 	}
 	
 	// バグ対策の、強制的な変更（隠蔽のため、このgetterは削除すること）
-	public BaseAttackController BAC
+	public AC.BaseAttackController BAC
 	{
 		get{ return _bac; }
 	}
